+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.14] - 2021-07-XX
+
+### Added
+
+- `proxyGet` method
+- `proxyGetFilter` method
+- `virtualProxyGet` method
+- `virtualProxyGetFilter` method
+- `virtualProxyRemove` method
+- `proxyMetadataExport` method
+
 ## [0.0.13] - 2021-07-25
 
 ### Added